backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTX==1.3.0
python-dotenv==1.0.1
gunicorn==21.2.0
PyMySQL==1.1.1
PyYAML==6.0.2
arxiv==2.1.3
openai>=1.0.0
requests>=2.31.0
pypdf>=4.0.0
